requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
marshmallow==3.19.0
python-dotenv==1.0.0
pytest==7.4.0
gunicorn==21.2.0 


# Google Services
google-generativeai==0.3.1
firebase-admin==6.2.0

# HTTP client
requests==2.31.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utils
pydantic==2.3.0
python-multipart==0.0.6
Pillow==10.0.0

# Added from the code block
jsonschema==4.19.0
Werkzeug==2.3.7